refactor(cc-addon-credentials): refactor api calls and create a `cc-addon-header.api.js` file
