+++ src/components/cc-addon-credentials-beta/cc-addon-credentials-beta.smart-keycloak.js
@@ -1,11 +1,9 @@
-import { sendToApi } from '../../lib/send-to-api.js';
-import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
-import '../cc-smart-container/cc-smart-container.js';
-// @ts-expect-error FIXME: remove when clever-client exports types
-import { get as getAddon } from '@clevercloud/client/esm/api/v2/addon.js';
 import { notifyError, notifySuccess } from '../../lib/notifications.js';
+import { defineSmartComponent } from '../../lib/smart/define-smart-component.js';
 import { generateDocsHref } from '../../lib/utils.js';
 import { i18n } from '../../translations/translation.js';
+import '../cc-smart-container/cc-smart-container.js';
+import { CcAddonCredentialsBetaApi } from './cc-addon-credentials-beta.api.js';
 import './cc-addon-credentials-beta.js';
 
 /** @type {AddonCredential[]} */
@@ -25,16 +23,16 @@ const SKELETON_DATA = [
     },
   },
 ];
+const PROVIDER_ID = 'keycloak';
 
 /**
  * @typedef {import('./cc-addon-credentials-beta.js').CcAddonCredentialsBeta} CcAddonCredentialsBeta
  * @typedef {import('./cc-addon-credentials-beta.types.js').AddonCredentialsBetaStateLoaded} AddonCredentialsBetaStateLoaded
- * @typedef {import('./cc-addon-credentials-beta.types.js').RawAddon} RawAddon
- * @typedef {import('./cc-addon-credentials-beta.types.js').KeycloakOperatorInfo} KeycloakOperatorInfo
  * @typedef {import('../cc-addon-credentials-content/cc-addon-credentials-content.types.js').AddonCredential} AddonCredential
  * @typedef {import('../cc-addon-credentials-content/cc-addon-credentials-content.types.js').AddonCredentialNg} AddonCredentialNg
  * @typedef {import('../cc-addon-credentials-content/cc-addon-credentials-content.types.js').AddonCredentialNgEnabled} AddonCredentialNgEnabled
  * @typedef {import('../cc-addon-credentials-content/cc-addon-credentials-content.types.js').AddonCredentialNgDisabled} AddonCredentialNgDisabled
+ * @typedef {import('../../operators.types.js').KeycloakOperatorInfo} KeycloakOperatorInfo
  * @typedef {import('../../lib/send-to-api.js').ApiConfig} ApiConfig
  * @typedef {import('../../lib/smart/smart-component.types.js').OnContextUpdateArgs<CcAddonCredentialsBeta>} OnContextUpdateArgs
  */
@@ -85,25 +83,16 @@ defineSmartComponent({
     });
 
     api
-      .getAddonWithOperator()
-      .then((operator) => {
+      .getCredentials()
+      .then((credentials) => {
         updateComponent('state', {
           type: 'loaded',
           tabs: {
-            default: [
-              {
-                code: 'user',
-                value: operator.envVars.CC_KEYCLOAK_ADMIN,
-              },
-              {
-                code: 'password',
-                value: operator.envVars.CC_KEYCLOAK_ADMIN_DEFAULT_PASSWORD,
-              },
-              {
-                code: 'ng',
-                value: formatNgData(operator.features.networkGroup),
-              },
-            ],
+            default: credentials,
+          },
+          docLink: {
+            text: 'Keycloak - Documentation',
+            href: 'https://www.clever-cloud.com/developers/doc/addons/keycloak/',
           },
         });
       })
@@ -196,7 +185,7 @@ function formatNgData(data) {
   };
 }
 
-class Api {
+class Api extends CcAddonCredentialsBetaApi {
   /**
    * @param {object} params
    * @param {ApiConfig} params.apiConfig
@@ -205,86 +194,27 @@ class Api {
    * @param {AbortSignal} params.signal
    */
   constructor({ apiConfig, ownerId, addonId, signal }) {
-    this._apiConfig = apiConfig;
-    this._ownerId = ownerId;
-    this._addonId = addonId;
-    this._provideId = null;
-    this._realId = null;
-    this._signal = signal;
-  }
-
-  /** @return {Promise<RawAddon>} */
-  _getAddon() {
-    return getAddon({ id: this._ownerId, addonId: this._addonId }).then(
-      sendToApi({ apiConfig: this._apiConfig, signal: this._signal }),
-    );
+    super({ apiConfig, ownerId, addonId, providerId: PROVIDER_ID, signal });
   }
 
   /**
-   * @param {string} realId
-   * @returns {Promise<KeycloakOperatorInfo>}
+   * @return {Promise<AddonCredential[]>}
    */
-  _getOperator(realId) {
-    return getOperator({ providerId: 'keycloak', realId }).then(
-      sendToApi({ apiConfig: this._apiConfig, signal: this._signal }),
-    );
+  async getCredentials() {
+    const operator = /** @type {KeycloakOperatorInfo} */ (await this.getAddonWithOperator());
+    return [
+      {
+        code: 'user',
+        value: operator.envVars.CC_KEYCLOAK_ADMIN,
+      },
+      {
+        code: 'password',
+        value: operator.envVars.CC_KEYCLOAK_ADMIN_DEFAULT_PASSWORD,
+      },
+      {
+        code: 'ng',
+        value: formatNgData(operator.features.networkGroup),
+      },
+    ];
   }
-
-  /** @returns {Promise<KeycloakOperatorInfo>} */
-  async getAddonWithOperator() {
-    const rawAddon = await this._getAddon();
-    const operator = await this._getOperator(rawAddon.realId);
-    this._realId = rawAddon.realId;
-    this._provideId = /** @type {import('../common.types.js').RawAddonProvider} */ (rawAddon.provider).id;
-
-    return operator;
-  }
-
-  createNg() {
-    return createNg({ providerId: this._provideId, realId: this._realId }).then(
-      sendToApi({ apiConfig: this._apiConfig }),
-    );
-  }
-
-  deleteNg() {
-    return deleteNg({ providerId: this._provideId, realId: this._realId }).then(
-      sendToApi({ apiConfig: this._apiConfig }),
-    );
-  }
-}
-
-/**
- * @param {{ providerId: string, realId: string }} params
- */
-function getOperator(params) {
-  // no multipath for /self or /organisations/{id}
-  return Promise.resolve({
-    method: 'get',
-    url: `/v4/addon-providers/addon-${params.providerId}/addons/${params.realId}`,
-    headers: { Accept: 'application/json' },
-    // no queryParams
-    // no body
-  });
-}
-
-/**
- * @param {{ providerId: string, realId: string }} params
- */
-function createNg({ providerId, realId }) {
-  return Promise.resolve({
-    method: 'post',
-    url: `/v4/addon-providers/addon-${providerId}/addons/${realId}/networkgroup`,
-    headers: { Access: 'application/json' },
-  });
-}
-
-/**
- * @param {{ providerId: string, realId: string }} params
- */
-function deleteNg({ providerId, realId }) {
-  return Promise.resolve({
-    method: 'delete',
-    url: `/v4/addon-providers/addon-${providerId}/addons/${realId}/networkgroup`,
-    headers: { Access: 'application/json' },
-  });
 }
